Correct assert values for integration test
The previous values used for the assert statement corresponded to results obtained from the code before the refactoring. In the commit `Introduce constant 'CONVERSION_STD_TO_FWHM'`, the precision of a conversion factor from standard deviation to FWHM was refined and used consistently throughout the code base, which led to a change in the results and is reflected in the updated values for the assert statements.